Add documentation, format the code, refactor, move code from turnstile.c to mutex.c
